Changed min/max value type from Double to Int

+++ Source/Peaks KYC/Models/ConfigModel.swift
@@ -38,5 +38,5 @@ struct ValidationConfig: Decodable {
     let regex: String?
     let message: String?
     let minLength, maxLength: Int?
-    let minValue, maxValue: Double?
+    let minValue, maxValue: Int?
 }
